+++ lib/logger.py
@@ -1,0 +1,10 @@
+"""Global logging helper"""
+import logging
+
+
+LOG_LEVEL = logging.DEBUG
+
+
+logging.basicConfig(level=LOG_LEVEL)
+
+log = logging.getLogger('upshot')